@@ -2,7 +2,7 @@ import 'package:dreamers/screens/auth_screen.dart';
 import 'package:dreamers/screens/dream_screen.dart';
 import 'package:dreamers/screens/favorite_screen.dart';
 import 'package:dreamers/screens/splash_screen.dart';
-import 'package:firebase_admob/firebase_admob.dart';
+
 
 import 'providers/auth_providers.dart';
 import 'screens/dream_detail_screen.dart';
@@ -14,8 +14,8 @@ import 'package:flutter/material.dart';
 import './providers/dreams_providers.dart';
 
 import 'package:firebase_core/firebase_core.dart';
-import 'package:firebase_database/firebase_database.dart';
-import 'package:firebase_database/ui/firebase_animated_list.dart';
+
+import 'package:google_mobile_ads/google_mobile_ads.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -32,23 +32,8 @@ Future<void> main() async {
     ),
   );*/
   runApp(DreamersApp());
-  myBanner
-    ..load()
-    ..show(
-      // Positions the banner ad 60 pixels from the bottom of the screen
-      anchorOffset: 0.0,
-      // Positions the banner ad 10 pixels from the center of the screen to the right
-      horizontalCenterOffset: 0.0,
-      // Banner Position
-      anchorType: AnchorType.bottom,
-    );
-  /* myInterstitial
-    ..load()
-    ..show(
-      anchorType: AnchorType.bottom,
-      anchorOffset: 0.0,
-      horizontalCenterOffset: 0.0,
-    );*/
+ 
+ 
 }
 
 class DreamersApp extends StatelessWidget {
@@ -87,7 +72,7 @@ class DreamersApp extends StatelessWidget {
                }})
             ,
             onGenerateRoute: (settings) {
-              print(settings.arguments);
+              //print(settings.arguments);
               return MaterialPageRoute(builder: (context) => HomePage());
             },
             onUnknownRoute: (settings) {
@@ -119,33 +104,3 @@ class DreamersApp extends StatelessWidget {
   }
 }
 
-MobileAdTargetingInfo targetingInfo = MobileAdTargetingInfo(
-  keywords: <String>['flutterio', 'beautiful apps'],
-  contentUrl: 'https://flutter.io',
-
-  childDirected: false,
-  testDevices: <String>[], // Android emulators are considered test devices
-);
-
-BannerAd myBanner = BannerAd(
-  // Replace the testAdUnitId with an ad unit id from the AdMob dash.
-  // https://developers.google.com/admob/android/test-ads
-  // https://developers.google.com/admob/ios/test-ads
-  adUnitId: BannerAd.testAdUnitId,
-  size: AdSize.smartBanner,
-  targetingInfo: targetingInfo,
-  listener: (MobileAdEvent event) {
-    print("BannerAd event is $event");
-  },
-);
-/*
-InterstitialAd myInterstitial = InterstitialAd(
-  // Replace the testAdUnitId with an ad unit id from the AdMob dash.
-  // https://developers.google.com/admob/android/test-ads
-  // https://developers.google.com/admob/ios/test-ads
-  adUnitId: InterstitialAd.testAdUnitId,
-  targetingInfo: targetingInfo,
-  listener: (MobileAdEvent event) {
-    print("InterstitialAd event is $event");
-  },
-);*/
